Moved prefer_headers.
